Add files via upload
In this project, you have to make a Font Manipulator Web Application. This means this web application can change the font size of a text by using left wrist and right wrist. Closer the left wrist and right wrist to each other smaller the font size, farther the left wrist and right wrist to each other bigger the font size. And this text will be nothing but your name.

For this project, you will have to design the UI for this webpage.